# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  handlers.AddProgramRequest:
    properties:
      ch:
        description: 頻道 ID
        example: channel_id
        type: string
      desc:
        description: 節目描述
        example: 節目描述
        type: string
      duration:
        description: 時長（秒）
        example: 300
        type: integer
      name:
        description: 節目名稱
        example: 節目名稱
        type: string
      tags:
        description: 標籤列表
        items:
          type: integer
        type: array
      updateCover:
        description: 是否更新頻道封面
        example: false
        type: boolean
      youtube_id:
        description: YouTube 影片 ID
        example: dQw4w9WgXcQ
        type: string
    required:
    - ch
    - name
    - youtube_id
    type: object
...
  handlers.DeleteProgramRequest:
    properties:
      ch:
        type: string
      ids:
        items:
          type: integer
        type: array
    required:
    - ch
    - ids
    type: object
...
  handlers.MoveProgramRequest:
    properties:
      ch:
        type: string
      ids:
        items:
          type: integer
        type: array
      target:
        type: string
    required:
    - ch
    - ids
    - target
    type: object
...
  handlers.SaveProgramOrderRequest:
    properties:
      ch:
        type: string
      order:
        items:
          type: integer
        type: array
    required:
    - ch
    - order
    type: object
  handlers.SaveProgramRequest:
    properties:
      ch:
        type: string
      desc:
        type: string
      duration:
        type: integer
      name:
        type: string
      prog_id:
        type: integer
      tags:
        items:
          type: integer
        type: array
      updateCover:
        type: boolean
      youtube_id:
        type: string
    required:
    - ch
    - name
    - prog_id
    - youtube_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@higgstv.com
    name: API Support
    url: http://www.higgstv.com
  description: Session-based authentication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HiggsTV API Server
  version: "1.0"
paths:
...
  /apis/addprog:
    post:
      consumes:
      - application/json
      description: 在頻道中新增節目（需要登入且為頻道管理員）
      parameters:
      - description: 新增節目請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 需要登入或權限不足" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 新增節目
      tags:
      - 節目
...
  /apis/delprog:
    post:
      consumes:
      - application/json
      description: 刪除節目（可一次刪除多個，需登入，需有權限）
      parameters:
      - description: 刪除節目請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DeleteProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 權限不足" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 刪除節目
      tags:
      - 節目
...
  /apis/pickprog:
    get:
      description: 透過 Bookmarklet 新增 YouTube 影片到未分類頻道（支援 JSONP），同時支援 youtube_id 和
        url 參數
      parameters:
      - description: JSONP callback 函數名
        in: query
        name: callback
        type: string
      - description: 影片名稱
        in: query
        name: name
        required: true
        type: string
      - description: YouTube 影片 ID（文件規範）
        in: query
        name: youtube_id
        type: string
      - description: YouTube URL（向後相容，與 youtube_id 二選一）
        in: query
        name: url
        type: string
      - description: 描述
        in: query
        name: desc
        type: string
      - description: 時長（秒）
        in: query
        name: duration
        type: integer
      - description: 標籤（逗號分隔的數字）
        in: query
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 需要登入或參數無效
      security:
      - ApiAuth: []
      summary: Pick 節目（Bookmarklet）
      tags:
      - 節目
  /apis/prog/saveorder:
    post:
      consumes:
      - application/json
      description: 儲存節目在頻道中的自訂排序（需登入，需有權限）
      parameters:
      - description: 儲存節目順序請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SaveProgramOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 權限不足" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 儲存節目順序
      tags:
      - 節目
  /apis/progmoveto:
    post:
      consumes:
      - application/json
      description: 將節目從一個頻道搬到另一個頻道（需登入，需有來源和目標頻道權限）
      parameters:
      - description: 移動節目請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.MoveProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 權限不足" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 移動節目
      tags:
      - 節目
...
  /apis/saveprog:
    post:
      consumes:
      - application/json
      description: 更新節目內容（需登入，需有權限）
      parameters:
      - description: 儲存節目請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SaveProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 權限不足" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 儲存節目
      tags:
      - 節目
...
swagger: "2.0"